requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
celery[redis]>=5.4.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
jsonschema>=4.21.0
anthropic>=0.40.0
aiofiles>=23.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
